Implemented filter for module types(barchart).
Implemented the possibility show only content from a specific moodle  module  type in the barchart tab.  E.g. if only the accesses on wikis should be displayed, the option wiki can be selected from the selectbox on the right.
These choices can of course be reverted to show all data.