Restructured most of the core files.
A bug that led to the linechart data missing some datasets was fixed.
Most of the code is now run by the javascript files and no longer on the server. Reason for this is that the restructuring of data done on the serverside had to be implemented on the clientside anyways to be able to filter data.
While restructuring the files like this, a lot (but not all) redundancies were removed in both the javascript and php files.
Prepared structure to be able to implement filters for data more easily.