All chartdata now mergeable and barchart data can now be filtered.
Previously downloaded data from all chart types can now be merged together. It has to be noted, that the timespan of the merged data is currently not displayed, it is therefore not possible for the user to check for gaps in the data.
Another change is, that the barchart can now be filtered, similar to the linechart and heatmap. It is also possible to filter data of only one type of activity/ressource by first selecting the type in the dropdown menu and then applying the filter.